2.86

+++ src/main/java/jp/co/toshiba/ppok/entity/City.java
@@ -32,7 +32,7 @@ public final class City implements Serializable {
 	 * This field corresponds to the database column ID
 	 */
 	@Id
-	private Long id;
+	private Integer id;
 
 	/**
 	 * This field corresponds to the database column NAME
@@ -56,7 +56,7 @@ public final class City implements Serializable {
 	 * This field corresponds to the database column POPULATION
 	 */
 	@Column(nullable = false)
-	private Long population;
+	private Integer population;
 
 	/**
 	 * This field corresponds to the database column DELETE_FLG
